Refined note auditioning behavior during drag and placement
- Prevented note auditioning while playback is on.
- When dragging multiple notes, audition only a single representative note.
- Cap the maximum audition duration to 1.5 seconds when placing a new note (previously it played the full note length).
- Add a slight delay to vertical movement auditioning when `keySnap` is disabled, preventing rapid-fire triggering.

+++ plugin/source/editor/panels/NotePathManager.cpp
@@ -11,7 +11,7 @@ void NotePathManager::addArcAsCubic(juce::Path &path, float centreX, float centr
 
     // 1. Split into segments of maximum 90 degrees (pi / 2)
     const float maxSegmentAngle = juce::MathConstants<float>::halfPi;
-    int numSegments = std::ceil(std::abs(totalAngle) / maxSegmentAngle);
+    int numSegments = static_cast<int>(std::ceil(std::abs(totalAngle) / maxSegmentAngle));
     if (numSegments < 1)
         numSegments = 1;
